+++ pyo3-derive-backend/src/args.rs
@@ -180,6 +180,7 @@ mod test {
 
     use args::{parse_arguments, Argument};
     use syn;
+    use proc_macro::TokenStream;
 
     fn items(s: TokenStream) -> Vec<syn::NestedMeta> {
         let dummy: syn::ItemFn = parse_quote!{#s fn dummy() {}};